+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2021-02-11
 ### Changed
 - Use Span<T>.Fill
+- CeilPow2 AggressiveInlining
 
 ## [1.2.7] - 2021-02-09
 ### Changed